// apps/mybase/src/task/type.ts
import { z } from "zod";

export const TaskSchema = z.object({
  id: z.string(),
  title: z.string(),
  detail: z.string(),
  dueDate: z.string().regex(/^\d{4}-\d{2}-\d{2}$/),
  priority: z.enum(["Low", "Medium", "High", "Extra-high"]),
  tag: z.array(z.string()),
  status: z.enum(["Todo", "Pending", "In-Progress", "Done"]),
  createdAt: z.string(),
});

export const TaskSchemaCreate = TaskSchema.omit({
  id: true,
  createdAt: true,
});

export type Task = z.infer<typeof TaskSchema>;

export type TaskCreate = z.infer<typeof TaskSchemaCreate>;